ajouter des films

+++ back/models/userMovie.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Column, Integer, String, ForeignKey
+from sqlalchemy import Column, Integer, String, ForeignKey, Float
 from database import Base
 
 class UserMovie(Base):
@@ -8,5 +8,5 @@ class UserMovie(Base):
     user_id = Column(Integer, ForeignKey("Users.id"), nullable=False)
     movie_id = Column(Integer, nullable=False)
     movie_img = Column(String(255))
-    movie_rating = Column(Integer)
+    movie_rating = Column(Float)
     movie_name = Column(String(255))
